fix(eval): serialize audit schema OID as a JSON number

+++ scripts/evaluation/audit_scope.sql
@@ -1,7 +1,7 @@
 SELECT pg_catalog.json_build_object(
  'database', pg_catalog.current_database(),
  'schema', :'schema',
- 'schema_oid', (SELECT oid FROM pg_catalog.pg_namespace WHERE nspname=:'schema'),
+ 'schema_oid', (SELECT oid::bigint FROM pg_catalog.pg_namespace WHERE nspname=:'schema'),
  'current_user', current_user, 'session_user', session_user,
  'read_only', pg_catalog.current_setting('transaction_read_only')='on',
  'select_allowed', (SELECT bool_and(pg_catalog.has_table_privilege(current_user,pg_catalog.format('%I.%I',:'schema',name),'SELECT')) FROM (VALUES ('runs'),('model_attempts'),('quota_reservations'),('artifacts'),('run_events')) AS required(name)),
